Resolve Comments Moved some function's
