+++ ocs_ci/utility/localstorage.py
@@ -134,6 +134,7 @@ def check_local_volume_set():
 
     Returns:
         bool: True if LVS present, False if LVS not present
+
     """
 
     if csv.get_csvs_start_with_prefix(
